Added POST /add/notes & turned off /resources caching

+++ src/main/java/com/gft/ResourceUtils.java
@@ -44,28 +44,26 @@ public class ResourceUtils {
     return contents;
   }
 
-  public boolean writeContentsToResource(String title, List<String> contents, String path) throws IOException, Exception {
-    File file = null;
-System.out.println("enter");
-System.out.println(title+contents.size()+path);
-    for (int i = 0; i < files.length; i++) {
-      if (files[i].getFilename().replaceAll(".md$", "").equals(title)) file = files[i].getFile();
+  public boolean writeContentsToResource(String title, List<String> contents, String path, boolean newFile) throws IOException, Exception {
+    if (!newFile) {
+      File file = null;
+
+      for (int i = 0; i < files.length; i++) {
+        if (files[i].getFilename().replaceAll(".md$", "").equals(title)) file = files[i].getFile();
+      }
+
+      if (file == null) throw new Exception("Couldn't find the given resource");
+      file.delete();
     }
-System.out.println(file);
-    if (file == null) throw new Exception("Couldn't find the given resource");
-    file.delete();
 
     ClassLoader classLoader = getClass().getClassLoader();
     File newf = new File(classLoader.getResource(".").getFile() + path);
-    System.out.println(newf);
     if (!newf.createNewFile()) {
       throw new Exception("Couldn't write to the the given resource");
     }
 
     FileWriter f2 = new FileWriter(newf, false);
     Optional<String> opt = contents.stream().reduce((tot, line) -> tot + "\n" + line);
-    System.out.println(opt);
-    System.out.println(opt.get());
     f2.write(opt.get());
     f2.close();
